Use the `secrecy` crate
This crate contained some internal types/traits which provide the same
functionality as the `secrecy` crate.

This commit removes those types/traits and switches to the equivalent
`secrecy` types/traits, most notably the `secrecy::Secret` wrapper type
(which ensures zeroization on drop) and the `ExposeSecret` for accessing
the inner secret value.

With this change, this crate no longer needs to depend directly on
`zeroize`, but uses it vicariously through `secrecy`.